refactor(inventory): add weighted products support and events

+++ Modules/InventoryModule/Inventory.Application/Dtos/ProductDtos/GetProductStockLocationDto.cs
@@ -6,7 +6,7 @@ namespace Inventory.Application.Dtos.ProductDtos
         public string LocationName { get; set; }
         public Guid WarehouseId { get; set; }
         public string WarehouseName { get; set; }
-        public int Quantity { get; set; }
-        public int ReservedQuantity { get; set; }
+        public decimal Quantity { get; set; }
+        public decimal ReservedQuantity { get; set; }
     }
 }
