+++ Modules/InventoryModule/Inventory.Application/Dtos/StockAdjustmentDtos/GetStockAdjustmentDto.cs
@@ -3,8 +3,8 @@ namespace Inventory.Application.Dtos.StockAdjustmentDtos
     public class GetStockAdjustmentDto
     {
         public Guid Id { get; set; }
-        public int ExpectedQuantity { get; set; }
-        public int ActualQuantity { get; set; }
+        public decimal ExpectedQuantity { get; set; }
+        public decimal ActualQuantity { get; set; }
         public DateTime Date { get; set; }
         public Guid UserId { get; set; }
         public Guid WarehouseId { get; set; }
