+++ Modules/InventoryModule/Inventory.Domain/Entities/StockQuant.cs
@@ -8,8 +8,8 @@ namespace Inventory.Domain.Entities
 {
     public class StockQuant : BaseEntity
     {
-        public int Quantity { get; set; }
-        public int ReservedQuantity { get; set; }
+        public decimal Quantity { get; set; }
+        public decimal ReservedQuantity { get; set; }
 
         public Guid ProductId { get; set; }
         public Product Product { get; set; }
